requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
